+++ src/movies/dto/base-movie-detail-response.dto.ts
@@ -1,0 +1,13 @@
+import { ReviewDto } from "./review.dto";
+
+export class BaseMovieDetailResponseDto {
+  id: number;
+  title: string;
+  overview: string;
+  posterPath: string;
+  voteAverage: number;
+  voteCount: number;
+  providers: string[];
+  theMovieDbId: number;
+  reviews: ReviewDto[];
+}
