+++ src/movies/dto/review.dto.ts
@@ -1,0 +1,5 @@
+export class ReviewDto {
+    id: number;
+    content: string;
+    createdAt: string; // ISO 8601 형식의 문자열
+}
